Move screenshare to settings folder
