+++ apps/wallet-mobile/src/InitApp.tsx
@@ -7,7 +7,7 @@ import * as Sentry from 'sentry-expo'
 import uuid from 'uuid'
 
 import {AppNavigator} from './AppNavigator'
-import {useInitScreenShare} from './features/ScreenShare'
+import {useInitScreenShare} from './features/Settings/ScreenShare'
 import {CONFIG, isProduction} from './legacy/config'
 import {useCrashReportsEnabled} from './yoroi-wallets/hooks'
 import {walletManager} from './yoroi-wallets/walletManager'
