+++ apps/wallet-mobile/src/features/Settings/ScreenShare/index.ts
@@ -1,0 +1,6 @@
+export {
+  useAllowScreenshots,
+  useChangeScreenShareSetting,
+  useInitScreenShare,
+  useScreenShareSettingEnabled,
+} from './ScreenShare'
