+++ apps/wallet-mobile/translations/messages/src/Catalyst/QrCode.json
@@ -6,12 +6,12 @@
     "start": {
       "line": 92,
       "column": 12,
-      "index": 2905
+      "index": 2914
     },
     "end": {
       "line": 95,
       "column": 3,
-      "index": 3003
+      "index": 3012
     }
   },
   {
@@ -21,12 +21,12 @@
     "start": {
       "line": 96,
       "column": 15,
-      "index": 3020
+      "index": 3029
     },
     "end": {
       "line": 99,
       "column": 3,
-      "index": 3142
+      "index": 3151
     }
   },
   {
@@ -36,12 +36,12 @@
     "start": {
       "line": 100,
       "column": 16,
-      "index": 3160
+      "index": 3169
     },
     "end": {
       "line": 105,
       "column": 3,
-      "index": 3390
+      "index": 3399
     }
   },
   {
@@ -51,12 +51,12 @@
     "start": {
       "line": 106,
       "column": 16,
-      "index": 3408
+      "index": 3417
     },
     "end": {
       "line": 111,
       "column": 3,
-      "index": 3626
+      "index": 3635
     }
   },
   {
@@ -66,12 +66,12 @@
     "start": {
       "line": 112,
       "column": 8,
-      "index": 3636
+      "index": 3645
     },
     "end": {
       "line": 115,
       "column": 3,
-      "index": 3784
+      "index": 3793
     }
   },
   {
@@ -81,12 +81,12 @@
     "start": {
       "line": 116,
       "column": 14,
-      "index": 3800
+      "index": 3809
     },
     "end": {
       "line": 119,
       "column": 3,
-      "index": 3891
+      "index": 3900
     }
   }
 ]
