+++ apps/wallet-mobile/translations/messages/src/features/Settings/ApplicationSettings/ApplicationSettingsScreen.json
@@ -6,12 +6,12 @@
     "start": {
       "line": 231,
       "column": 11,
-      "index": 8067
+      "index": 8064
     },
     "end": {
       "line": 234,
       "column": 3,
-      "index": 8177
+      "index": 8174
     }
   },
   {
@@ -21,12 +21,12 @@
     "start": {
       "line": 235,
       "column": 21,
-      "index": 8200
+      "index": 8197
     },
     "end": {
       "line": 238,
       "column": 3,
-      "index": 8333
+      "index": 8330
     }
   },
   {
@@ -36,12 +36,12 @@
     "start": {
       "line": 239,
       "column": 18,
-      "index": 8353
+      "index": 8350
     },
     "end": {
       "line": 242,
       "column": 3,
-      "index": 8465
+      "index": 8462
     }
   },
   {
@@ -51,12 +51,12 @@
     "start": {
       "line": 243,
       "column": 22,
-      "index": 8489
+      "index": 8486
     },
     "end": {
       "line": 246,
       "column": 3,
-      "index": 8610
+      "index": 8607
     }
   },
   {
@@ -66,12 +66,12 @@
     "start": {
       "line": 247,
       "column": 9,
-      "index": 8621
+      "index": 8618
     },
     "end": {
       "line": 250,
       "column": 3,
-      "index": 8721
+      "index": 8718
     }
   },
   {
@@ -81,12 +81,12 @@
     "start": {
       "line": 251,
       "column": 18,
-      "index": 8741
+      "index": 8738
     },
     "end": {
       "line": 254,
       "column": 3,
-      "index": 8868
+      "index": 8865
     }
   },
   {
@@ -96,12 +96,12 @@
     "start": {
       "line": 255,
       "column": 13,
-      "index": 8883
+      "index": 8880
     },
     "end": {
       "line": 258,
       "column": 3,
-      "index": 8992
+      "index": 8989
     }
   },
   {
@@ -111,12 +111,12 @@
     "start": {
       "line": 259,
       "column": 15,
-      "index": 9009
+      "index": 9006
     },
     "end": {
       "line": 262,
       "column": 3,
-      "index": 9122
+      "index": 9119
     }
   },
   {
@@ -126,12 +126,12 @@
     "start": {
       "line": 263,
       "column": 19,
-      "index": 9143
+      "index": 9140
     },
     "end": {
       "line": 266,
       "column": 3,
-      "index": 9304
+      "index": 9301
     }
   },
   {
@@ -141,12 +141,12 @@
     "start": {
       "line": 267,
       "column": 20,
-      "index": 9326
+      "index": 9323
     },
     "end": {
       "line": 270,
       "column": 3,
-      "index": 9460
+      "index": 9457
     }
   },
   {
@@ -156,12 +156,12 @@
     "start": {
       "line": 271,
       "column": 24,
-      "index": 9486
+      "index": 9483
     },
     "end": {
       "line": 274,
       "column": 3,
-      "index": 9669
+      "index": 9666
     }
   },
   {
@@ -171,12 +171,12 @@
     "start": {
       "line": 275,
       "column": 18,
-      "index": 9689
+      "index": 9686
     },
     "end": {
       "line": 278,
       "column": 3,
-      "index": 9820
+      "index": 9817
     }
   },
   {
@@ -186,12 +186,12 @@
     "start": {
       "line": 279,
       "column": 22,
-      "index": 9844
+      "index": 9841
     },
     "end": {
       "line": 282,
       "column": 3,
-      "index": 10025
+      "index": 10022
     }
   },
   {
@@ -201,12 +201,12 @@
     "start": {
       "line": 283,
       "column": 13,
-      "index": 10040
+      "index": 10037
     },
     "end": {
       "line": 286,
       "column": 3,
-      "index": 10148
+      "index": 10145
     }
   },
   {
@@ -216,12 +216,12 @@
     "start": {
       "line": 287,
       "column": 17,
-      "index": 10167
+      "index": 10164
     },
     "end": {
       "line": 290,
       "column": 3,
-      "index": 10284
+      "index": 10281
     }
   }
 ]
